组件调整

+++ SecCircuitDiagramPlugin/tpseccircuitdiagram.h
@@ -10,7 +10,7 @@ class QDESIGNER_WIDGET_EXPORT TPSecCircuitDiagram : public QWidget
 
     Q_PROPERTY(int chargeLevel READ chargeLevel WRITE setChargeLevel NOTIFY chargeLevelChanged)
     Q_PROPERTY(int warningLevel READ warningLevel WRITE setWarningLevel NOTIFY warningLevelChanged)
-    Q_PROPERTY(int systemVoltage READ systemVoltage WRITE setSystemVoltage NOTIFY systemVoltageChanged)
+    Q_PROPERTY(double systemVoltage READ systemVoltage WRITE setSystemVoltage NOTIFY systemVoltageChanged)
     Q_PROPERTY(bool mainContactorClosed READ mainContactorClosed WRITE setMainContactorClosed NOTIFY mainContactorClosedChanged)
     Q_PROPERTY(bool dischargeContactorClosed READ dischargeContactorClosed WRITE setDischargeContactorClosed NOTIFY dischargeContactorClosedChanged)
     Q_PROPERTY(bool chargeContactorClosed READ chargeContactorClosed WRITE setChargeContactorClosed NOTIFY chargeContactorClosedChanged)
@@ -35,8 +35,8 @@ public:
     int warningLevel() const;
     void setWarningLevel(int level);
 
-    int systemVoltage() const;
-    void setSystemVoltage(int voltage);
+    double systemVoltage() const;
+    void setSystemVoltage(double voltage);
 
     bool mainContactorClosed() const;
     void setMainContactorClosed(bool closed);
@@ -129,7 +129,7 @@ protected:
 private:
     int m_chargeLevel;
     int m_warningLevel;
-    int m_systemVoltage;
+    double m_systemVoltage;
     bool m_mainContactorClosed;
     bool m_dischargeContactorClosed;
     bool m_chargeContactorClosed;
